api_chat/requirements.txt
# requirements.txt:
fastapi==0.115.9
uvicorn==0.34.2
langchain==0.3.24
langchain-openai==0.3.14
langchain-community==0.3.22
chromadb==1.0.6
python-dotenv==1.1.0
